% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eucDis}
\alias{eucDis}
\title{Euclidean Distance (eucDis)}
\usage{
eucDis(rawData1, rawData2)
}
\arguments{
\item{rawData1}{A vector containing the first of the two sets of time series data.}

\item{rawData2}{A vector containing the second of the two sets of time series data.}
}
\value{
The euclidean distance between the two sets of time series data as a float.
}
\description{
Compute the euclidean distance between two sets of time series data.
}

