% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minDis}
\alias{minDis}
\title{Minimum Distance (minDis)}
\usage{
minDis(SAXData1, card1, SAXData2, card2, rawDataSize,
  suppressWarnings = FALSE)
}
\arguments{
\item{card1}{A vector containing the first of two sets of cardinality data for the SAX data.}

\item{card2}{A vector containing the second of two sets of cardinality data for the SAX data.}

\item{rawDataSize}{The size of the time series data before conversion to SAX data. The size of each time series data should be the same for both SAX data sets.}

\item{suppressWarnings}{Supress any warnings created by the user. Not any warnings created by the compiler}

\item{SaxData1}{A vector containing the first of two sets of SAX data for minimum distance computation.}

\item{SaxData2}{A vector containing the second of two sets of SAX data for minimum distance computation.}
}
\value{
The minimum distance between the two SAX data sets as a float
}
\description{
Compute the minimum distance between two sets of SAX data.
}

