% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct Data (reconstruct)}
\usage{
reconstruct(seq, card, step, centroidList)
}
\arguments{
\item{seq}{A vector of SAX data to be reconstructed using centroids.}

\item{card}{A vector of cardinality values for the SAX data.}

\item{step}{The size of the distance between indices of sampling of the SAX data.}

\item{centroidList}{A list of lists of centroids and their cardinalities. Computed from 'runKMedian' function.}
}
\value{
A list containing a vector storing the reconstructed data and a vector storing the difference vector between the original and reconstructed data.
}
\description{
Reconstruct a sequence of SAX data using centroids trained using 'runKMedian' function and compute the difference vector between the original and reconstructed data.
}

