% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runKMedian}
\alias{runKMedian}
\title{kMedian Centroids (runKMedian)}
\usage{
runKMedian(seq, card, step, windowSize, clusterNum, centroidSeq, centroidCard,
  prevCentroid = FALSE)
}
\arguments{
\item{seq}{A matrix of SAX data to be used for computation of centroids. Each row representing a new/seperate piece of data from the same system.}

\item{card}{A matrix of cardinality values for the SAX data. This should have the same dimensions as the paraketer 'seq'.}

\item{step}{The size of the distance between indices of sampling of the SAX data.}

\item{windowSize}{The size of each sample from the SAX data. Also the number of elements in each centroid.}

\item{clusterNum}{The number of centroids generated.}

\item{centroidSeq}{A matrix of centroids only used for computation if 'prevCentroid' parameter is true.}

\item{centroidCard}{A matrix of cardinalities of centroids only used for computation if 'prevCentroid' parameter is true.}

\item{prevCentroid}{False: Do not use inputted centroids in computation. True: Use inputted centroids in computation.}
}
\value{
A list of lists of computed centroids and their cardinalities.
}
\description{
Compute the centroids of a matrix of SAX data using kMedians.
}

