% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runSAX}
\alias{runSAX}
\title{Convert to SAX (runSAX)}
\usage{
runSAX(orgData, segmentSize, alphabetSize, iSAX = FALSE)
}
\arguments{
\item{orgData}{A vector of time series data to be converted to a SAX Word.}

\item{segmentSize}{The size of each segment taken when converting to SAX Word.}

\item{alphabetSize}{The size of the alphabet for classification of PAAs to SAX Words.}

\item{iSAX}{False: returns only the SAX Word. True: returns SAX Word and Cardinality.}
}
\value{
A Vector that contains the SAX Word and if iSAX = true a Vector that contains its corresponding Cardinality.
}
\description{
Convert a time series into a SAX Word.
}

